fix mke2fs option

+++ tools/update_prebuilt.sh
@@ -11,15 +11,25 @@ cp -f $2/idbloader.img $1/
 cp -f $2/misc.img $1/
 cp -f $2/dtbo.img $1/
 
+TOP=$PWD
+export MKE2FS_CONFIG="${TOP}/tools/mke2fs.conf"
+if [ ! -f ${MKE2FS_CONFIG} ]; then
+    echo "error: ${MKE2FS_CONFIG} not found."
+    exit 1
+fi
+true ${MKFS:="${TOP}/tools/mke2fs"}
 
+RET=0
 if [ ! -f $1/userdata.img ]; then
 	USERDATA_SIZE=104857600
 	echo "Generating empty userdata.img (size:${USERDATA_SIZE})"
 	TMPDIR=`mktemp -d`
     IMG_BLK=$((${USERDATA_SIZE} / 4096))
     [ -f $1/userdata.img ] && rm -f $1/userdata.img
-    MKE2FS_CONFIG=${TOP}/tools/mke2fs.conf ${PWD}/tools/mke2fs -E android_sparse -t ext4 -L userdata -M /root -b 4096 -d ${TMPDIR} $1/userdata.img ${IMG_BLK}
+	INODE_SIZE=256
+    ${MKFS} -N ${INODE_SIZE} -E android_sparse -t ext4 -L userdata -M /root -b 4096 -d ${TMPDIR} $1/userdata.img ${IMG_BLK}
+	RET=$?
 	rm -rf ${TMPDIR}
 fi
 
-exit $?
+exit $RET
